refactor: de-part game_region_minimap cluster into libraries (#4117)
Convert the 6-file GameRegionMinimap part cluster to explicit-import
libraries (painter, visible panel, zoom chrome). Shrink
repo.app_no_part_directives grandfather by six entries; behavior unchanged.

Refs #4117
